refactor(tui): split prompt input screen modules
